concurrent wave snapshot A: k_fieldRuntime reconstruction + bpc_main, camp menu/persona, datCalendarPrelude
One of four divergent snapshots of the same change id, produced by
parallel agents sharing a single jj working copy. Carries the shared
k_fieldRuntime.c reconstruction (+468) plus bpc_main.c rework and
incidental edits to _h_camp_menu, _h_camp_persona, k_fldFilter,
k_data, k_window, op_fade_mid, datCalendarPrelude, fmGslCont, h_pad
and nmSequence. Description added retroactively; content unchanged.

+++ src/nmSequence.c
@@ -48,7 +48,7 @@ u64 FUN_003c57e0(u64 param_1,int param_2);
 u64 FUN_003c5810(u64 param_1,int param_2);
 void FUN_003c5840(u64 param_1,int param_2);
 u64 FUN_003c5870(u64 param_1,int param_2);
-u32 FUN_003c58f0(u32 param_1,u32 param_2,u32 param_3,u32 param_4);
+u64 FUN_003c58f0(u32 param_1,long param_2,u32 param_3,u32 param_4);
 void FUN_003c5a20(void);
 u32 FUN_003c5a40(u32 param_1,u32 param_2,u32 param_3,u32 param_4);
 void FUN_003c5af0(u32 param_1);
@@ -68,7 +68,7 @@ void FUN_003c5d40(u32 param_1,int param_2,int param_3);
 #define FUN_003c5810(...) ((u64 (*)(...))FUN_003c5810)(__VA_ARGS__)
 #define FUN_003c5840(...) ((void (*)(...))FUN_003c5840)(__VA_ARGS__)
 #define FUN_003c5870(...) ((u64 (*)(...))FUN_003c5870)(__VA_ARGS__)
-#define FUN_003c58f0(...) ((u32 (*)(...))FUN_003c58f0)(__VA_ARGS__)
+#define FUN_003c58f0(...) ((u64 (*)(...))FUN_003c58f0)(__VA_ARGS__)
 #define FUN_003c5a20(...) ((void (*)(...))FUN_003c5a20)(__VA_ARGS__)
 #define FUN_003c5a40(...) ((u32 (*)(...))FUN_003c5a40)(__VA_ARGS__)
 #define FUN_003c5af0(...) ((void (*)(...))FUN_003c5af0)(__VA_ARGS__)
@@ -581,7 +581,7 @@ void FUN_003c58e0(void)
 // FUN_003C58F0 NONMATCHING
 
 
-u32 FUN_003c58f0(u32 param_1,u32 param_2,u32 param_3,u32 param_4)
+u64 FUN_003c58f0(u32 param_1,long param_2,u32 param_3,u32 param_4)
